Remove handle from list of things to retweet
The problem with retweeting everything using the bot handle
is that we end up retweeting entire threads, most of which are
totally irrelevant.

This commit removes the bot handle from the list of terms that
are retweeted, so that we do not end up retweeting these
(in all likelihood irrelevant) tweets.